+++ R/read-parquet.R
@@ -2,7 +2,12 @@
 #'
 #' Converts the contents of the named Parquet file to a R data frame.
 #'
-#' @param file Path to a Parquet file.
+#' @param file Path to a Parquet file. It may also be an R connection,
+#'   in which case it first reads all data from the connection, writes
+#'   it into a temporary file, then reads the temporary file, and
+#'   deletes it. The connection might be open, it which case it must be
+#'   a binary connection. If it is not open, then `read_parquet()` will
+#'   open it and also close it in the end.
 #' @param options Nanoparquet options, see [parquet_options()].
 #' @return A `data.frame` with the file's contents.
 #' @export
@@ -17,6 +22,11 @@
 #' print(str(parquet_df))
 
 read_parquet <- function(file, options = parquet_options()) {
+	if (inherits(file, "connection")) {
+		tmp <- tempfile(fileext = ".parquet")
+		dump_connection(file, tmp)
+		file <- tmp
+	}
   file <- path.expand(file)
   res <- .Call(nanoparquet_read2, file, options, sys.call())
 	dicts <- res[[2]]
@@ -45,4 +55,23 @@ read_parquet <- function(file, options = parquet_options()) {
 	}
 
   res
+}
+
+# dump the contents of a connection to path
+dump_connection <- function(con, path) {
+	if (!isOpen(con)) {
+		on.exit(close(con), add = TRUE)
+		open(con, "rb")
+	}
+	ocon <- file(path, open = "wb")
+	# 10 MB buffer by default
+	bs <- getOption("nanoparquet.con_buffer_size", 1024L * 1024L * 10)
+	while (TRUE) {
+		buf <- readBin(con, what = "raw", n = bs)
+		if (length(buf) == 0) {
+			break
+		}
+		writeBin(buf, path)
+	}
+	close(ocon)
 }