+++ NEWS.md
@@ -53,6 +53,8 @@
 * `write_parquet(file = ":raw:")` now works correctly for larger data
   frames (#77).
 
+* `read_parquet()` can now read from an R connection (#71).
+
 * `read_parquet()` now reads `DECIMAL` values correctly from `INT32`
   and `INT64` columns if their `scale` is not zero.